projects/agents/pyproject.toml
[project]
name = "athanor-agents"
version = "0.1.0"
description = "Athanor AI agent framework — LangGraph agents with vLLM backend"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg[binary]>=3.2",
    "langchain-openai>=0.3",
    "langchain-core>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "ddgs>=7.0",
    "redis>=5.0",
    "pymupdf>=1.25",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "fastembed>=0.7",
    "PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
